requirements.txt
# torch==1.10.2
torchdiffeq==0.2.3
numpy==1.25.0
pandas==2.1.1
scipy==1.11.3
scikit-learn==1.3.1
matplotlib==3.8.0
seaborn==0.11.0
tqdm==4.66.1